restructuring -> msg related objects are now in Msg subdirectory; deleted depreacted files

+++ R_Visualizer/Msg/msg.cpp
@@ -5,7 +5,7 @@ Msg::Msg()
     this->timestamp = QDateTime();
     this->id = 0x0u;
     this->data = {
-         0,0,0,0,0,0,0,0
+         static_cast<quint8>(0),static_cast<quint8>(0),static_cast<quint8>(0),static_cast<quint8>(0),static_cast<quint8>(0),static_cast<quint8>(0),static_cast<quint8>(0),static_cast<quint8>(0)
     };
 }
 
@@ -17,7 +17,14 @@ Msg::Msg(QDateTime timestamp, unsigned int id, MsgDataT &data) : timestamp(times
 Msg::Msg(QDateTime timestamp, unsigned int id, QByteArray &data) : timestamp(timestamp), id(id)
 {
     this->data = {
-         data.at(0),0,0,0,0,0,0,0
+        static_cast<quint8>(data.at(0)),
+        static_cast<quint8>(0),
+        static_cast<quint8>(0),
+        static_cast<quint8>(0),
+        static_cast<quint8>(0),
+        static_cast<quint8>(0),
+        static_cast<quint8>(0),
+        static_cast<quint8>(0)
     };
     if(data.length() > 1 )
     {
@@ -79,21 +86,13 @@ void Msg::setId(unsigned int value)
 
 quint8 Msg::getCode() const
 {
-    //return this->data.at(0x7u);
     return this->data.code;
 }
 
 void Msg::setCode(quint8 value)
 {
-    //this->data.at(0x7u) = value;
     this->data.code = value;
 }
-
-//ToDO
-//unsigned int Msg::getMessage() const
-//{
-//    return message;
-//}
 
 void Msg::setMessage(quint8 data0, quint8 data1, quint8 data2, quint8 data3, quint8 data4, quint8 data5, quint8 data6)
 {
@@ -130,7 +129,16 @@ void Msg::parseIN(QJsonObject jsonMsg)
     this->timestamp = QDateTime::fromString(jsonMsg["timestamp"].toString(),QString("dd.MM.yyyy - hh:mm:ss.zzz"));
     this->id = jsonMsg["id"].toInt();
     QJsonObject jsonData = jsonMsg["data"].toObject();
-    this->data  = { jsonData["code"].toInt(), jsonData["data0"].toInt(), jsonData["data1"].toInt(), jsonData["data2"].toInt(), jsonData["data3"].toInt(), jsonData["data4"].toInt(), jsonData["data5"].toInt(), jsonData["data6"].toInt() };
+    this->data  = {
+        static_cast<quint8>(jsonData["code"].toInt()),
+        static_cast<quint8>(jsonData["data0"].toInt()),
+        static_cast<quint8>(jsonData["data1"].toInt()),
+        static_cast<quint8>(jsonData["data2"].toInt()),
+        static_cast<quint8>(jsonData["data3"].toInt()),
+        static_cast<quint8>(jsonData["data4"].toInt()),
+        static_cast<quint8>(jsonData["data5"].toInt()),
+        static_cast<quint8>(jsonData["data6"].toInt())
+    };
 }
 
 QJsonObject Msg::parseOUT() const
